refactor: extract enums exclude and comment
