+++ frontend/svelte/src/tests/lib/utils/proposals.utils.spec.ts
@@ -1,9 +1,6 @@
 import type { ProposalInfo } from "@dfinity/nns";
-import { Topic } from "@dfinity/nns";
-import { enumKeys, enumValues } from "../../../lib/utils/enum.utils";
 import {
   emptyProposals,
-  excludeTopics,
   lastProposalId,
 } from "../../../lib/utils/proposals.utils";
 
@@ -28,29 +25,4 @@ describe("proposals-utils", () => {
     expect(lastProposalId(mockProposals)).toEqual(
       mockProposals[mockProposals.length - 1].id
     ));
-
-  it("should exclude all topics", () =>
-    expect(excludeTopics([])).toEqual(enumValues(Topic)));
-
-  it("should exclude no topics", () => {
-    const all: Topic[] = enumKeys(Topic).map((key: string) => Topic[key]);
-
-    expect(excludeTopics(all)).toEqual([]);
-  });
-
-  it("should exclude selected topics", () => {
-    const results: Topic[] = [
-      Topic.Unspecified,
-      Topic.ManageNeuron,
-      Topic.ExchangeRate,
-      Topic.NetworkEconomics,
-      Topic.Governance,
-      Topic.NodeAdmin,
-      Topic.ParticipantManagement,
-      Topic.NetworkCanisterManagement,
-      Topic.NodeProviderRewards,
-    ];
-
-    expect(excludeTopics([Topic.SubnetManagement, Topic.Kyc])).toEqual(results);
-  });
 });
